+++ packages/ui/react/package.json
@@ -11,7 +11,6 @@
     "lint": "eslint \"**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}\""
   },
   "dependencies": {
-    "@favolink-ui/hooks": "workspace:*",
     "@favolink-ui/icons": "workspace:*",
     "@vanilla-extract/css": "^1.14.1",
     "@vanilla-extract/dynamic": "^2.1.0",
