+++ packages/ui/react/src/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export * from '@favolink-ui/hooks';
 export * from '@favolink-ui/icons';
-export * from './system';
+export * from './hooks';
 export * from './styles';
+export * from './system';
 export * from './utils';
 export * from './components';
